Move Feature back to ui.map as it is not used outside of ui layer
