+++ app/src/main/java/org/groundplatform/android/ui/datacollection/tasks/point/DropPinTaskMapFragment.kt
@@ -19,8 +19,8 @@ import androidx.lifecycle.LiveData
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
 import org.groundplatform.android.model.map.CameraPosition
-import org.groundplatform.android.model.map.Feature
 import org.groundplatform.android.ui.datacollection.tasks.AbstractTaskMapFragment
+import org.groundplatform.android.ui.map.Feature
 import org.groundplatform.android.ui.map.MapFragment
 
 @AndroidEntryPoint
