+++ app/src/main/java/org/groundplatform/android/ui/datacollection/tasks/polygon/DrawAreaTaskMapFragment.kt
@@ -26,8 +26,8 @@ import kotlinx.coroutines.flow.combine
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.launch
 import org.groundplatform.android.model.map.CameraPosition
-import org.groundplatform.android.model.map.Feature
 import org.groundplatform.android.ui.datacollection.tasks.AbstractTaskMapFragment
+import org.groundplatform.android.ui.map.Feature
 import org.groundplatform.android.ui.map.gms.GmsExt.toBounds
 
 @AndroidEntryPoint
