+++ app/src/main/java/org/groundplatform/android/ui/map/gms/features/FeatureClusterManager.kt
@@ -23,7 +23,7 @@ import com.google.maps.android.clustering.ClusterManager
 import com.google.maps.android.collections.MarkerManager
 import org.groundplatform.android.R
 import org.groundplatform.android.model.map.Bounds
-import org.groundplatform.android.model.map.Feature
+import org.groundplatform.android.ui.map.Feature
 import org.groundplatform.android.ui.map.gms.GmsExt.toBounds
 import org.groundplatform.android.ui.map.gms.toGoogleMapsObject
