+++ app/src/main/java/org/groundplatform/android/ui/map/gms/features/FeatureClusterRenderer.kt
@@ -24,8 +24,8 @@ import com.google.maps.android.clustering.Cluster
 import com.google.maps.android.clustering.ClusterManager
 import com.google.maps.android.clustering.view.DefaultClusterRenderer
 import org.groundplatform.android.common.Constants.CLUSTERING_ZOOM_THRESHOLD
-import org.groundplatform.android.model.map.Feature
 import org.groundplatform.android.ui.IconFactory
+import org.groundplatform.android.ui.map.Feature
 import org.groundplatform.android.ui.map.gms.CLUSTER_Z
 
 /**
