+++ app/src/main/java/org/groundplatform/android/ui/map/gms/features/LineStringRenderer.kt
@@ -26,7 +26,7 @@ import com.google.android.gms.maps.model.PolylineOptions
 import javax.inject.Inject
 import org.groundplatform.android.R
 import org.groundplatform.android.model.geometry.LineString
-import org.groundplatform.android.model.map.Feature
+import org.groundplatform.android.ui.map.Feature
 import org.groundplatform.android.ui.map.gms.POLYLINE_Z
 import org.groundplatform.android.ui.map.gms.toLatLngList
 import org.groundplatform.android.ui.util.BitmapUtil
