+++ app/src/main/java/org/groundplatform/android/ui/map/gms/features/MapsItemManager.kt
@@ -26,7 +26,7 @@ import org.groundplatform.android.model.geometry.LineString
 import org.groundplatform.android.model.geometry.MultiPolygon
 import org.groundplatform.android.model.geometry.Point
 import org.groundplatform.android.model.geometry.Polygon
-import org.groundplatform.android.model.map.Feature
+import org.groundplatform.android.ui.map.Feature
 
 /** Manages [Feature]s displayed on the map as Maps SDK items (marker, polyline, etc). */
 class MapsItemManager(
